Component DailyReports was completed.

+++ FinManagerWebClient/Library/Report.cs
@@ -5,6 +5,6 @@ using FinManagerWebClient.Model;
 public abstract class Report
 {
     public List<TransactionVM> TransactionsVM { get; set; }
-    public int TotalIncome { get; set; }
-    public int TotalExpenses { get; set; }
+    public double TotalIncome { get; set; }
+    public double TotalExpenses { get; set; }
 }
